+++ packages/assets/src/wallets/index.ts
@@ -13,3 +13,4 @@ export * from './trust';
 export * from './mobile-wallet';
 export * from './sui-wallet';
 export * from './suiet';
+export * from './solflare';