+++ packages/web3/src/connect-modal/components/WalletList.tsx
@@ -1,20 +1,27 @@
-import React, { forwardRef, useContext, useImperativeHandle, useMemo } from 'react';
-import { QrcodeOutlined } from '@ant-design/icons';
-import { Button, Empty, List, Space, Typography } from 'antd';
-import classNames from 'classnames';
+import {
+  forwardRef,
+  ForwardRefRenderFunction,
+  useContext,
+  useImperativeHandle,
+  useMemo,
+} from 'react';
+import { Empty, List } from 'antd';
+import mobile from 'is-mobile';
 
 import { connectModalContext } from '../context';
 import type { ConnectModalActionType, ConnectModalProps, Wallet } from '../interface';
 import { defaultGroupOrder } from '../utils';
-import PluginTag from './PluginTag';
-import WalletIcon from './WalletIcon';
+import WalletItem from './WalletItem';
 
 export type WalletListProps = Pick<
   ConnectModalProps,
   'walletList' | 'group' | 'groupOrder' | 'emptyProps'
 >;
 
-const WalletList = forwardRef<ConnectModalActionType, WalletListProps>((props, ref) => {
+const WalletList: ForwardRefRenderFunction<ConnectModalActionType, WalletListProps> = (
+  props,
+  ref,
+) => {
   const { walletList = [], group: internalGroup, groupOrder, emptyProps } = props;
   const { prefixCls, updateSelectedWallet, selectedWallet, localeMessage, updatePanelRoute } =
     useContext(connectModalContext);
@@ -44,6 +51,15 @@ const WalletList = forwardRef<ConnectModalActionType, WalletListProps>((props, r
   const needGrouping =
     internalGroup !== false && (internalGroup !== undefined || groupKeys.length > 1);
 
+  const openInUniversalLink = (wallet: Wallet) => {
+    const url = wallet.deeplink?.urlTemplate
+      .replace('${url}', encodeURIComponent(window.location.href))
+      .replace('${ref}', encodeURIComponent(window.location.href));
+    if (url) {
+      window.location.href = url;
+    }
+  };
+
   const selectWallet = async (wallet: Wallet) => {
     const hasWalletReady = await wallet.hasWalletReady?.();
     if (hasWalletReady) {
@@ -53,6 +69,9 @@ const WalletList = forwardRef<ConnectModalActionType, WalletListProps>((props, r
         updateSelectedWallet(wallet, {
           connectType: 'extension',
         });
+      } else if (mobile()) {
+        // open in universal link
+        openInUniversalLink(wallet);
       } else if (wallet.getQrCode) {
         // Extension not installed and can use qr code to connect
         updateSelectedWallet(wallet, {
@@ -64,11 +83,15 @@ const WalletList = forwardRef<ConnectModalActionType, WalletListProps>((props, r
       }
       return;
     }
-
-    // wallet not ready
-    // go to wallet page
-    updateSelectedWallet(wallet);
-    updatePanelRoute('wallet', true);
+    if (mobile() && wallet.deeplink) {
+      // open in universal link
+      openInUniversalLink(wallet);
+    } else {
+      // wallet not ready
+      // go to wallet page
+      updateSelectedWallet(wallet);
+      updatePanelRoute('wallet', true);
+    }
   };
 
   useImperativeHandle(ref, () => {
@@ -84,45 +107,18 @@ const WalletList = forwardRef<ConnectModalActionType, WalletListProps>((props, r
         dataSource={group ? dataSource[group] : walletList}
         rowKey="key"
         renderItem={(item) => (
-          <List.Item
-            className={classNames(`${prefixCls}-wallet-item`, {
-              selected:
-                item.key !== undefined
-                  ? selectedWallet?.key === item.key
-                  : selectedWallet?.name === item.name,
-            })}
-            onClick={() => {
-              selectWallet(item);
+          <WalletItem
+            wallet={item}
+            prefixCls={prefixCls}
+            selectedWallet={selectedWallet}
+            onSelect={selectWallet}
+            onQrCodeSelect={(wallet) => {
+              updateSelectedWallet(wallet, {
+                connectType: 'qrCode',
+              });
             }}
-          >
-            <div className={`${prefixCls}-content`}>
-              <WalletIcon wallet={item} />
-              <Typography.Text ellipsis={{ tooltip: true }} className={`${prefixCls}-name`}>
-                {item.name}
-              </Typography.Text>
-            </div>
-            <Space>
-              <PluginTag wallet={item} />
-              {item.getQrCode ? (
-                <Button
-                  size="small"
-                  className={`${prefixCls}-qr-btn`}
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    updateSelectedWallet(item, {
-                      connectType: 'qrCode',
-                    });
-                  }}
-                >
-                  <QrcodeOutlined />
-                </Button>
-              ) : (
-                walletList.some((w) => w.getQrCode) && (
-                  <div className={`${prefixCls}-qr-icon-empty`} />
-                )
-              )}
-            </Space>
-          </List.Item>
+            showQrPlaceholder={walletList.some((w) => w.getQrCode && w.hasExtensionInstalled)}
+          />
         )}
       />
     );
@@ -160,6 +156,14 @@ const WalletList = forwardRef<ConnectModalActionType, WalletListProps>((props, r
       )}
     </div>
   );
-});
+};
 
-export default WalletList;
+type WalletListComponent = {
+  (props: WalletListProps & { ref?: React.Ref<ConnectModalActionType> }): React.ReactNode;
+  displayName?: string;
+};
+
+const ForwardRefWalletList: WalletListComponent = forwardRef(WalletList);
+ForwardRefWalletList.displayName = 'WalletList';
+
+export default ForwardRefWalletList;