+++ packages/web3/src/solana/demos/more-wallets.tsx
@@ -1,10 +1,18 @@
 import React from 'react';
 import { ConnectButton, Connector } from '@ant-design/web3';
-import { CoinbaseWallet, PhantomWallet, SolanaWeb3ConfigProvider } from '@ant-design/web3-solana';
+import {
+  CoinbaseWallet,
+  OKXWallet,
+  PhantomWallet,
+  SolanaWeb3ConfigProvider,
+  SolflareWallet,
+} from '@ant-design/web3-solana';
 
 const App: React.FC = () => {
   return (
-    <SolanaWeb3ConfigProvider wallets={[CoinbaseWallet(), PhantomWallet()]}>
+    <SolanaWeb3ConfigProvider
+      wallets={[CoinbaseWallet(), PhantomWallet(), SolflareWallet(), OKXWallet()]}
+    >
       <Connector>
         <ConnectButton />
       </Connector>