+++ packages/web3/src/solana/demos/wallet-connect.tsx
@@ -1,15 +1,11 @@
 import React from 'react';
 import { ConnectButton, Connector } from '@ant-design/web3';
-import {
-  PhantomWallet,
-  SolanaWeb3ConfigProvider,
-  WalletConnectWallet,
-} from '@ant-design/web3-solana';
+import { SolanaWeb3ConfigProvider, WalletConnectWallet } from '@ant-design/web3-solana';
 
 const App: React.FC = () => {
   return (
     <SolanaWeb3ConfigProvider
-      wallets={[PhantomWallet(), WalletConnectWallet()]}
+      wallets={[WalletConnectWallet()]}
       walletConnect={{
         projectId: YOUR_WALLET_CONNECT_PROJECT_ID,
       }}