+++ packages/solana/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ant-design/web3-solana
 
+## 1.3.0
+
+### Minor Changes
+
+- b6f1f1d: feat: support SolflareWallet
+
+### Patch Changes
+
+- Updated dependencies [b6f1f1d]
+  - @ant-design/web3-assets@1.12.0
+  - @ant-design/web3-common@1.18.0
+
 ## 1.2.7
 
 ### Patch Changes