+++ packages/wagmi/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ant-design/web3-wagmi
 
+## 2.10.2
+
+### Patch Changes
+
+- Updated dependencies [b6f1f1d]
+  - @ant-design/web3-assets@1.12.0
+  - @ant-design/web3-common@1.18.0
+
 ## 2.10.1
 
 ### Patch Changes